+++ _templates/plugin/new/tsconfig.json.t
@@ -2,17 +2,9 @@
 to: packages/plugin-<%= name %>/tsconfig.json
 ---
 {
-    "compilerOptions": {
-        "target": "es5",
-        "module": "commonjs",
-        "declaration": true,
-        "outDir": "./dist",
-        "strict": true,
-        "esModuleInterop": true
-    },
-    "exclude": [
-        "node_modules",
-        "dist/",
-        "test/"
-    ]
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "lib",
+    "outDir": "dist"
+  }
 }