+++ packages/plugin-amplitude-node/lib/index.ts
@@ -1,26 +1,28 @@
 /* eslint-disable no-unused-vars, class-methods-use-this */
 import {
-  ItlyOptions,
-  ItlyEvent,
-  ItlyProperties,
-  ItlyPluginBase,
+  Options,
+  Event,
+  Properties,
+  PluginBase,
 } from '@itly/sdk-node';
 import Amplitude, { AmplitudeOptions } from 'amplitude';
 
 export { AmplitudeOptions };
 
-export default class AmplitudeNodePlugin extends ItlyPluginBase {
+export default class AmplitudeNodePlugin extends PluginBase {
   static ID: string = 'amplitude';
 
-  private amplitude: Amplitude;
+  private amplitude?: Amplitude;
 
-  constructor(apiKey: string, options?: AmplitudeOptions) {
+  constructor(
+    private apiKey: string,
+    private options?: AmplitudeOptions,
+  ) {
     super();
-    this.amplitude = new Amplitude(apiKey, options);
   }
 
   // TODO: Allow passing in an instance rather than adding
-  // TODO: destination specific configuration to ItlyOptions
+  // TODO: destination specific configuration to Options
   // constructor(instanceOrApiKey: Amplitude | string, options?: AmplitudeOptions) {
   //     if(typeof instanceOrApiKey === 'string') {
   //         this.amplitude = new Amplitude(instanceOrApiKey, options);
@@ -32,15 +34,19 @@ export default class AmplitudeNodePlugin extends ItlyPluginBase {
 
   id = () => AmplitudeNodePlugin.ID;
 
-  identify(userId: string, properties?: ItlyProperties) {
-    this.amplitude.identify({
+  load() {
+    this.amplitude = new Amplitude(this.apiKey, this.options);
+  }
+
+  identify(userId: string, properties?: Properties) {
+    this.amplitude!.identify({
       user_id: userId,
       user_properties: properties,
     });
   }
 
-  track(userId: string, event: ItlyEvent) {
-    this.amplitude.track({
+  track(userId: string, event: Event) {
+    this.amplitude!.track({
       event_type: event.name,
       user_id: userId,
       event_properties: event.properties,