+++ packages/plugin-mixpanel-node/lib/index.ts
@@ -1,36 +1,42 @@
 /* eslint-disable no-unused-vars, class-methods-use-this, import/no-unresolved */
 import {
-  ItlyPluginBase, ItlyEvent, ItlyProperties,
+  PluginBase, Event, Properties,
 } from '@itly/sdk-node';
 import Mixpanel, { InitConfig } from 'mixpanel';
 
 export interface MixpanelOptions extends InitConfig {}
 
-export default class MixpanelNodePlugin extends ItlyPluginBase {
+export default class MixpanelNodePlugin extends PluginBase {
   static ID: string = 'mixpanel';
 
-  private mixpanel: Mixpanel.Mixpanel;
+  private mixpanel?: Mixpanel.Mixpanel;
 
-  constructor(apiKey: string, mixpanelOptions: MixpanelOptions) {
+  constructor(
+    private apiKey: string,
+    private options?: MixpanelOptions,
+  ) {
     super();
-    this.mixpanel = Mixpanel.init(apiKey, mixpanelOptions);
   }
 
   id = () => MixpanelNodePlugin.ID;
 
-  alias(userId: string, previousId: string) {
-    this.mixpanel.alias(previousId, userId);
+  load() {
+    this.mixpanel = Mixpanel.init(this.apiKey, this.options);
   }
 
-  identify(userId: string, properties: ItlyProperties | undefined) {
-    this.mixpanel.people.set(userId, {
+  alias(userId: string, previousId: string) {
+    this.mixpanel!.alias(previousId, userId);
+  }
+
+  identify(userId: string, properties: Properties | undefined) {
+    this.mixpanel!.people.set(userId, {
       distinct_id: userId,
       ...properties,
     });
   }
 
-  track(userId: string, event: ItlyEvent) {
-    this.mixpanel.track(event.name, {
+  track(userId: string, event: Event) {
+    this.mixpanel!.track(event.name, {
       distinct_id: userId,
       ...event.properties,
     });