+++ packages/plugin-schema-validator/lib/index.ts
@@ -2,44 +2,69 @@
 /* eslint-disable no-unused-vars, class-methods-use-this */
 import Ajv from 'ajv';
 import {
-  ItlyEvent,
-  ItlyPluginBase,
+  Event,
+  PluginBase,
   ValidationResponse,
 } from '@itly/sdk-core';
 
 export type ValidationResponseHandler = (
   validation: ValidationResponse,
-  event: ItlyEvent,
+  event: Event,
   schema: any
 ) => any;
 
-const DEFAULT_VALIDATION_RESPONSE_HANDLER: ValidationResponseHandler = () => {};
+export type SchemaMap = { [schemaKey: string]: any };
 
-export default class SchemaValidatorPlugin extends ItlyPluginBase {
+const SYSTEM_EVENTS = ['identify', 'context', 'group', 'page'];
+function isSystemEvent(name: string) {
+  return SYSTEM_EVENTS.includes(name);
+}
+
+function isEmpty(obj: any) {
+  return obj === undefined || Object.keys(obj).length === 0;
+}
+
+export default class SchemaValidatorPlugin extends PluginBase {
   static ID: string = 'schema-validator';
 
-  private schemas: { [id: string]: any };
+  private ajv?: Ajv.Ajv;
 
-  private validators: { [id: string]: any };
+  private validators?: { [schemaKey: string]: Ajv.ValidateFunction };
 
-  private ajv: Ajv.Ajv;
-
-  private validationErrorHandler: ValidationResponseHandler;
-
-  constructor(schemas: { [id: string]: any }, validationErrorHandler?: ValidationResponseHandler) {
+  constructor(
+    private schemas: SchemaMap,
+    private validationErrorHandler: ValidationResponseHandler = () => {},
+  ) {
     super();
-    this.schemas = schemas;
-    this.ajv = new Ajv();
-    this.validators = {};
-    this.validationErrorHandler = validationErrorHandler || DEFAULT_VALIDATION_RESPONSE_HANDLER;
   }
 
   id = () => SchemaValidatorPlugin.ID;
 
-  validate(event: ItlyEvent): ValidationResponse {
+  load() {
+    this.ajv = new Ajv();
+    this.validators = {};
+  }
+
+  validate(event: Event): ValidationResponse {
     const schemaKey = this.getSchemaKey(event);
     // Check that we have a schema for this event
     if (!this.schemas[schemaKey]) {
+      if (isSystemEvent(schemaKey)) {
+        // pass system events by default
+        if (isEmpty(event.properties)) {
+          return {
+            valid: true,
+            pluginId: this.id(),
+          };
+        }
+
+        return {
+          valid: false,
+          message: `'${event.name}' schema is empty but properties were found. properties=${JSON.stringify(event.properties)}`,
+          pluginId: this.id(),
+        };
+      }
+
       return {
         valid: false,
         message: `Event ${event.name} not found in tracking plan.`,
@@ -48,23 +73,26 @@ export default class SchemaValidatorPlugin extends ItlyPluginBase {
     }
 
     // Compile validator for this event if needed
-    if (!this.validators[schemaKey]) {
-      this.validators[schemaKey] = this.ajv.compile(this.schemas[schemaKey]);
+    const validators = this.validators!;
+    if (!validators[schemaKey]) {
+      validators[schemaKey] = this.ajv!.compile(this.schemas[schemaKey]);
     }
 
-    const validator = this.validators[schemaKey];
+    const validator = validators[schemaKey]!;
     if (event.properties && !(validator(event.properties) === true)) {
-      const errors = validator.errors.map((e: any) => {
-        let extra = '';
-        if (e.keyword === 'additionalProperties') {
-          extra = ` (${e.params.additionalProperty})`;
-        }
-        return `\`properties${e.dataPath}\` ${e.message}${extra}.`;
-      }).join(' ');
+      const errorMessage = validator.errors
+        ? validator.errors.map((e: any) => {
+          let extra = '';
+          if (e.keyword === 'additionalProperties') {
+            extra = ` (${e.params.additionalProperty})`;
+          }
+          return `\`properties${e.dataPath}\` ${e.message}${extra}.`;
+        }).join(' ')
+        : 'An unknown error occurred during validation.';
 
       return {
         valid: false,
-        message: `Passed in ${event.name} properties did not validate against your tracking plan. ${errors}`,
+        message: `Passed in ${event.name} properties did not validate against your tracking plan. ${errorMessage}`,
         pluginId: this.id(),
       };
     }
@@ -75,12 +103,12 @@ export default class SchemaValidatorPlugin extends ItlyPluginBase {
     };
   }
 
-  validationError(validation: ValidationResponse, event: ItlyEvent) {
+  validationError(validation: ValidationResponse, event: Event) {
     const schemaKey = this.getSchemaKey(event);
     this.validationErrorHandler(validation, event, this.schemas[schemaKey]);
   }
 
-  getSchemaKey(event: ItlyEvent) {
+  getSchemaKey(event: Event) {
     return event.name;
   }
 }