+++ packages/plugin-segment-node/lib/index.ts
@@ -1,7 +1,7 @@
 
 /* eslint-disable no-unused-vars, class-methods-use-this */
 import {
-  ItlyPluginBase, ItlyEvent, ItlyProperties,
+  PluginBase, Event, Properties,
 } from '@itly/sdk-node';
 import Segment from 'analytics-node';
 
@@ -12,39 +12,45 @@ export type SegmentOptions = {
   enable?: boolean; // (default: true)
 }
 
-export default class SegmentNodePlugin extends ItlyPluginBase {
+export default class SegmentNodePlugin extends PluginBase {
   static ID = 'segment';
 
-  private segment: Segment;
+  private segment?: Segment;
 
-  constructor(apiKey: string, segmentOptions?: SegmentOptions) {
+  constructor(
+    private writeKey: string,
+    private options?: SegmentOptions,
+  ) {
     super();
-    this.segment = new Segment(apiKey, segmentOptions);
   }
 
   id = () => SegmentNodePlugin.ID;
 
-  alias(userId: string, previousId: string) {
-    this.segment.alias({ userId, previousId });
+  load() {
+    this.segment = new Segment(this.writeKey, this.options);
   }
 
-  identify(userId: string, properties: ItlyProperties | undefined) {
-    this.segment.identify({
+  alias(userId: string, previousId: string) {
+    this.segment!.alias({ userId, previousId });
+  }
+
+  identify(userId: string, properties: Properties | undefined) {
+    this.segment!.identify({
       userId,
       traits: { ...properties },
     });
   }
 
-  group(userId: string, groupId: string, properties: ItlyProperties | undefined) {
-    this.segment.group({
+  group(userId: string, groupId: string, properties: Properties | undefined) {
+    this.segment!.group({
       userId,
       groupId,
       traits: properties,
     });
   }
 
-  page(userId: string, category: string, name: string, properties: ItlyProperties | undefined) {
-    this.segment.page({
+  page(userId: string, category: string, name: string, properties: Properties | undefined) {
+    this.segment!.page({
       userId,
       category,
       name,
@@ -52,8 +58,8 @@ export default class SegmentNodePlugin extends ItlyPluginBase {
     });
   }
 
-  track(userId: string, event: ItlyEvent) {
-    this.segment.track({
+  track(userId: string, event: Event) {
+    this.segment!.track({
       userId,
       event: event.name,
       properties: { ...event.properties },